pyproject.toml
[project]
name = "flutter-classifier"
version = "0.1.0"
description = ""
authors = [
    {name = "maximkes",email = "max_pechin@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "pyaml (>=25.1.0,<26.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "kagglehub (>=0.3.12,<0.4.0)",
    "sphinx (>=8.2.3,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
